+++ HISTORY.rst
@@ -2,7 +2,14 @@
 History
 =======
 
+0.1.1 (2021-01-02)
+------------------
+
+* Implemented Marginal likelihood GP and Conditional.
+* Added corresponding notebook.
+
+
 0.1.0 (2021-01-01)
 ------------------
 
-* First release on PyPI.
+* First release.